+++ src/pages/Exchange/MainArea/Trade/Chart/Styles.ts
@@ -13,9 +13,9 @@ const fadeIn = keyframes`
   }
 `;
 
-export const ChangeArrow = styled.span<{isPositive: boolean}>`
-  animation: ${({isPositive}) =>
-    isPositive
+export const ChangeArrow = styled.span<{$isPositive: boolean}>`
+  animation: ${({$isPositive}) =>
+    $isPositive
       ? css`
           ${fadeIn} 0.5s ease-out
         `
@@ -49,12 +49,12 @@ export const ChartHeader = styled.div`
   margin-bottom: 20px;
 `;
 
-export const ChartTypeButton = styled.button<{active: boolean}>`
+export const ChartTypeButton = styled.button<{$active: boolean}>`
   align-items: center;
-  background: ${({active}) => (active ? colors.palette.blue['100'] : 'transparent')};
+  background: ${({$active}) => ($active ? colors.palette.blue['100'] : 'transparent')};
   border: none;
   border-radius: 8px;
-  color: ${({active}) => (active ? colors.palette.blue['500'] : colors.palette.gray['500'])};
+  color: ${({$active}) => ($active ? colors.palette.blue['500'] : colors.palette.gray['500'])};
   cursor: pointer;
   display: flex;
   justify-content: center;
@@ -62,8 +62,8 @@ export const ChartTypeButton = styled.button<{active: boolean}>`
   transition: all 0.2s ease;
 
   &:hover {
-    background: ${({active}) => (active ? colors.palette.blue['100'] : colors.palette.gray['100'])};
-    color: ${({active}) => (active ? colors.palette.blue['500'] : colors.palette.gray['700'])};
+    background: ${({$active}) => ($active ? colors.palette.blue['100'] : colors.palette.gray['100'])};
+    color: ${({$active}) => ($active ? colors.palette.blue['500'] : colors.palette.gray['700'])};
   }
 
   svg {
@@ -109,11 +109,11 @@ export const PriceAmount = styled.span`
   font-weight: 700;
 `;
 
-export const PriceChange = styled.div<{isPositive: boolean}>`
+export const PriceChange = styled.div<{$isPositive: boolean}>`
   align-items: center;
-  background: ${({isPositive}) => (isPositive ? 'rgba(22, 170, 22, 0.1)' : 'rgba(220, 13, 22, 0.1)')};
+  background: ${({$isPositive}) => ($isPositive ? 'rgba(22, 170, 22, 0.1)' : 'rgba(220, 13, 22, 0.1)')};
   border-radius: 8px;
-  color: ${({isPositive}) => (isPositive ? colors.palette.green['500'] : colors.palette.red['500'])};
+  color: ${({$isPositive}) => ($isPositive ? colors.palette.green['500'] : colors.palette.red['500'])};
   display: inline-flex;
   font-size: 14px;
   font-weight: 600;
@@ -173,20 +173,20 @@ export const StatsBar = styled.div`
   padding: 8px 12px;
 `;
 
-export const TimeframeButton = styled.button<{active: boolean}>`
-  background: ${({active}) => (active ? colors.white : 'transparent')};
+export const TimeframeButton = styled.button<{$active: boolean}>`
+  background: ${({$active}) => ($active ? colors.white : 'transparent')};
   border: none;
   border-radius: 8px;
-  box-shadow: ${({active}) => (active ? '0 2px 8px rgba(0, 0, 0, 0.08)' : 'none')};
-  color: ${({active}) => (active ? colors.palette.darkGray['500'] : colors.palette.gray['600'])};
+  box-shadow: ${({$active}) => ($active ? '0 2px 8px rgba(0, 0, 0, 0.08)' : 'none')};
+  color: ${({$active}) => ($active ? colors.palette.darkGray['500'] : colors.palette.gray['600'])};
   cursor: pointer;
   font-size: 14px;
-  font-weight: ${({active}) => (active ? '600' : '500')};
+  font-weight: ${({$active}) => ($active ? '600' : '500')};
   padding: 8px 16px;
   transition: all 0.2s ease;
 
   &:hover {
-    background: ${({active}) => (active ? colors.white : 'rgba(255, 255, 255, 0.5)')};
+    background: ${({$active}) => ($active ? colors.white : 'rgba(255, 255, 255, 0.5)')};
     color: ${colors.palette.darkGray['500']};
   }
 `;
@@ -197,39 +197,4 @@ export const TimeframeButtons = styled.div`
   display: flex;
   gap: 4px;
   padding: 4px;
-`;
-
-export const TooltipContainer = styled.div`
-  animation: ${fadeIn} 0.2s ease-out;
-  background: rgba(15, 20, 25, 0.95);
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 12px;
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.2);
-  padding: 12px 16px;
-`;
-
-export const TooltipDate = styled.div`
-  color: ${colors.palette.gray['400']};
-  font-size: 12px;
-  letter-spacing: 0.05em;
-  margin-bottom: 8px;
-  text-transform: uppercase;
-`;
-
-export const TooltipLabel = styled.span`
-  color: ${colors.palette.gray['300']};
-  font-size: 12px;
-`;
-
-export const TooltipPrice = styled.div`
-  align-items: baseline;
-  display: flex;
-  gap: 8px;
-`;
-
-export const TooltipValue = styled.span`
-  color: ${colors.white};
-  font-size: 16px;
-  font-weight: 600;
 `;
